feat: add block with details for person and rout
