+++ src/components/PeopleList.tsx
@@ -1,3 +1,4 @@
+import { Link } from 'react-router-dom';
 import { PeopleListProps, Person } from '../interfaces';
 
 function PeopleList(props: PeopleListProps) {
@@ -6,15 +7,17 @@ function PeopleList(props: PeopleListProps) {
   }
   return (
     <div className="people-list">
-      {props.people?.map((person: Person, index: number) => (
-        <div key={index} className="person">
-          <i className="fa-solid fa-user"></i>
-          <div className="person__name">Name: {person.name}</div>
-          <div className="person__height">Height: {person.height}</div>
-          <div className="person__mass">Mass: {person.mass}</div>
-          <div className="person__height">Eye color: {person.eye_color}</div>
-          <div className="person__mass">Gender: {person.gender}</div>
-        </div>
+      {props.people?.map((person: Person) => (
+        <Link to={`details/${person.name}`}>
+          <div key={person.name} className="person">
+            <i className="fa-solid fa-user"></i>
+            <div className="person__name">Name: {person.name}</div>
+            <div className="person__height">Height: {person.height}</div>
+            <div className="person__mass">Mass: {person.mass}</div>
+            <div className="person__height">Eye color: {person.eye_color}</div>
+            <div className="person__mass">Gender: {person.gender}</div>
+          </div>
+        </Link>
       ))}
     </div>
   );
