+++ src/main.tsx
@@ -1,10 +1,26 @@
 import ReactDOM from 'react-dom/client';
-
 import React from 'react';
-import MainApp from './MainApp.tsx';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import ErrorPage from './ErrorPage';
+import Root from './routes/Root';
+import PersonDetails from './routes/PersonDetails';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Root />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: 'details/:detailsId',
+        element: <PersonDetails />,
+      },
+    ],
+  },
+]);
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <MainApp />
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
