+++ src/routes/PersonDetails.tsx
@@ -1,0 +1,7 @@
+export default function PersonDetails() {
+  return (
+    <div className="person-info">
+      <h3>Detail information about person</h3>
+    </div>
+  );
+}
